MockAi Backend App/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.0
cycler==0.11.0
distro==1.9.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
fonttools==4.42.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
install==1.3.5
kiwisolver==1.4.4
matplotlib==3.7.2
multidict==6.0.5
numpy==1.25.2
openai==0.28.0
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
